New connection pool
* Added Redis transactions support
* Added more flexible connection pool
